// src/notifications/enums/notification.enums.ts
export enum NotificationChannel {
  IN_APP = 'IN_APP',
  EMAIL = 'EMAIL',
  WHATSAPP = 'WHATSAPP',
  PUSH = 'PUSH',
}

export enum NotificationEventType {
  APPROVAL_PENDING = 'APPROVAL_PENDING',
  APPROVAL_APPROVED = 'APPROVAL_APPROVED',
  APPROVAL_REJECTED = 'APPROVAL_REJECTED',
  SEND_BACK = 'SEND_BACK',
  SLA_BREACH = 'SLA_BREACH',
  MATERIAL_EXCEPTION = 'MATERIAL_EXCEPTION',
  PO_APPROVAL = 'PO_APPROVAL',
  INVOICE_MISMATCH = 'INVOICE_MISMATCH',
  PAYMENT_STATUS = 'PAYMENT_STATUS',
  RENEWAL_REMINDER = 'RENEWAL_REMINDER',
  ESCALATED = 'ESCALATED',
  LEAD_ASSIGNED = 'LEAD_ASSIGNED',
  OPPORTUNITY_ASSIGNED = 'OPPORTUNITY_ASSIGNED',
  TENDER_ASSIGNED = 'TENDER_ASSIGNED',
  TENDER_SUBMITTED = 'TENDER_SUBMITTED',
  QUOTATION_SUBMITTED = 'QUOTATION_SUBMITTED',
  CLIENT_DECISION = 'CLIENT_DECISION',
  PROJECT_CREATED = 'PROJECT_CREATED',
  PROJECT_STATUS_CHANGED = 'PROJECT_STATUS_CHANGED',
  MATERIAL_REQUEST_SUBMITTED = 'MATERIAL_REQUEST_SUBMITTED',
  SITE_ACCESS_REQUESTED = 'SITE_ACCESS_REQUESTED',
  SITE_ACCESS_DECIDED = 'SITE_ACCESS_DECIDED',
  BOQ_SUBMITTED = 'BOQ_SUBMITTED',
  PO_APPROVAL_PENDING = 'PO_APPROVAL_PENDING',
  PO_APPROVED = 'PO_APPROVED',
  PO_REJECTED = 'PO_REJECTED',
  PO_SENT_TO_VENDOR = 'PO_SENT_TO_VENDOR',
  PO_ACKNOWLEDGED = 'PO_ACKNOWLEDGED',
  DELIVERY_SCHEDULED = 'DELIVERY_SCHEDULED',
  DELIVERY_IN_TRANSIT = 'DELIVERY_IN_TRANSIT',
  DELIVERY_RECEIVED = 'DELIVERY_RECEIVED',
  GRN_SUBMITTED = 'GRN_SUBMITTED',
  GRN_APPROVAL_PENDING = 'GRN_APPROVAL_PENDING',
  GRN_APPROVED = 'GRN_APPROVED',
  GRN_REJECTED = 'GRN_REJECTED',
  PO_PARTIALLY_DELIVERED = 'PO_PARTIALLY_DELIVERED',
  PO_FULLY_DELIVERED = 'PO_FULLY_DELIVERED',
  INVOICE_SUBMITTED = 'INVOICE_SUBMITTED',
  INVOICE_ACCOUNTS_REVIEW = 'INVOICE_ACCOUNTS_REVIEW',
  INVOICE_MATCH_COMPLETED = 'INVOICE_MATCH_COMPLETED',
  INVOICE_HOLD = 'INVOICE_HOLD',
  INVOICE_RELEASED = 'INVOICE_RELEASED',
  INVOICE_REJECTED = 'INVOICE_REJECTED',
  INVOICE_APPROVED = 'INVOICE_APPROVED',
}