// src/material-requests/enums/material-request.enums.ts
export enum MaterialRequestStatus {
  DRAFT = 'DRAFT',
  SUBMITTED = 'SUBMITTED',
  UNDER_REVIEW = 'UNDER_REVIEW',
  PENDING_APPROVAL = 'PENDING_APPROVAL',
  APPROVED = 'APPROVED',
  REJECTED = 'REJECTED',
  SEND_BACK = 'SEND_BACK',
  CANCELLED = 'CANCELLED',
  CLOSED = 'CLOSED',
}

export enum ExceptionStatus {
  NOT_CHECKED = 'NOT_CHECKED',
  NORMAL = 'NORMAL',
  EXCEPTION = 'EXCEPTION',
}